Friendlier appearance of quiz button
Changed from .btn-danger to .btn-success which should make it green.
I have not yet set up the site locally so I'm coding in the dark here,
also changed from .fa-danger to .fa-success but not sure if that
makes any difference. :)

I did this based on the code in the master branch, then realized that
there were other branches. It probably looks like shit and might not
even work, but you have to try, right? ;)