Friendlier appearance of quiz button on start page
Changed from .btn-danger to .btn-success which should make it green.
I have not yet set up the site locally so I'm coding in the dark here,
also changed from .fa-danger to .fa-success but not sure if that
makes any difference. :)

This was previously added to the user-pleasure branch, but I'm now
committing it to master directly after popular demand.